// src/objects/enums/AccessoryId.ts
export enum AccessoryId {
  I21110 = "21110",
  I21120 = "21120",
  I21210 = "21210",
  I21220 = "21220",
  I21310 = "21310",
  I21320 = "21320",

  I22201 = "22201",
  I22202 = "22202",
  I22203 = "22203",
  I22204 = "22204",
  I22205 = "22205",
  I22206 = "22206",
  I22207 = "22207",

  I22311 = "22311",
  I22312 = "22312",
  I22313 = "22313",
  I22314 = "22314",
  I22315 = "22315",
  I22316 = "22316",
  I22317 = "22317",
  I22321 = "22321",
  I22322 = "22322",
  I22323 = "22323",
  I22324 = "22324",
  I22325 = "22325",
  I22326 = "22326",
  I22327 = "22327",

  I22411 = "22411",
  I22412 = "22412",
  I22413 = "22413",
  I22414 = "22414",
  I22415 = "22415",
  I22416 = "22416",
  I22417 = "22417",
  I22421 = "22421",
  I22422 = "22422",
  I22423 = "22423",
  I22424 = "22424",
  I22425 = "22425",
  I22426 = "22426",
  I22427 = "22427",

  I23201 = "23201",
  I23202 = "23202",
  I23203 = "23203",
  I23204 = "23204",

  I23311 = "23311",
  I23312 = "23312",
  I23313 = "23313",
  I23314 = "23314",
  I23321 = "23321",
  I23322 = "23322",
  I23323 = "23323",
  I23324 = "23324",

  I23411 = "23411",
  I23412 = "23412",
  I23413 = "23413",
  I23414 = "23414",
  I23421 = "23421",
  I23422 = "23422",
  I23423 = "23423",
  I23424 = "23424",

  I34400 = "34400",
  I34401 = "34401",
  I34402 = "34402",
  I34403 = "34403",
  I34404 = "34404",
  I34405 = "34405",
  I34406 = "34406",
  I34407 = "34407",
  I34408 = "34408",
  I34409 = "34409",
  I34410 = "34410",
  I34411 = "34411",
  I34412 = "34412",
  I34413 = "34413",
  I34414 = "34414",
  I34415 = "34415",
  I34416 = "34416",

  I36001 = "36001",
  I36002 = "36002",
  I36003 = "36003",
  I36004 = "36004",
  I36005 = "36005",
  I36006 = "36006",
  I36007 = "36007",
  I36008 = "36008",
  I36009 = "36009",
  I36010 = "36010",
  I36011 = "36011",
  I36012 = "36012",
  I36013 = "36013",
  I36014 = "36014",
}